Move DummySpecialization to sc-network (#4680)
